requirements.txt
click==8.1.3
Flask==2.3.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.1
python-dotenv==1.0.0
scikit-learn==1.2.2
SQLAlchemy==2.0.12
tensorflow==2.12.0
openai==0.27.7
psycopg2==2.9.6